// gai-backend/src/modules/payments-expenses/domain/enums/payment-expense-audit-operation.enum.ts
export enum PaymentExpenseAuditOperation {
  CREATE_PAYMENT = 'CREATE_PAYMENT',
  UPDATE_PAYMENT = 'UPDATE_PAYMENT',
  APPROVE_PAYMENT = 'APPROVE_PAYMENT',
  MARK_PAYMENT_AS_PAID = 'MARK_PAYMENT_AS_PAID',
  CANCEL_PAYMENT = 'CANCEL_PAYMENT',
  CREATE_EXPENSE = 'CREATE_EXPENSE',
  UPDATE_EXPENSE = 'UPDATE_EXPENSE',
  APPROVE_EXPENSE = 'APPROVE_EXPENSE',
  REJECT_EXPENSE = 'REJECT_EXPENSE',
  MARK_EXPENSE_AS_PAID = 'MARK_EXPENSE_AS_PAID',
  CANCEL_EXPENSE = 'CANCEL_EXPENSE',
  CREATE_ATTACHMENT_UPLOAD_URL = 'CREATE_ATTACHMENT_UPLOAD_URL',
  CONFIRM_ATTACHMENT_UPLOAD = 'CONFIRM_ATTACHMENT_UPLOAD',
  DOWNLOAD_ATTACHMENT_URL = 'DOWNLOAD_ATTACHMENT_URL',
  REMOVE_ATTACHMENT = 'REMOVE_ATTACHMENT',
}